Recfactor is_valid_workout to use fewer network calls
This uses fewer network calls by caching the "canary cells" and passing
them into is_valid_workout as an argument instead of fetching the value
of each canary cell individually, and making a separate call to get all
the values for each sheet. If all three canary cells are blank, we can
assume the sheet is blank or at least worth skipping.